Annotation update

+++ annotations/nocedit.exe/pseudocode/include/types/structs/SRenderVertex.h
@@ -8,8 +8,8 @@
 // Ghidra size: 0x30 (48 bytes)
 typedef struct SRenderVertex {
     SProjectedVertex projected_vertex; // 0x0
-    float u; // 0x18
-    float v; // 0x1c
+    int u; // 0x18
+    int v; // 0x1c
     float light; // 0x20
     int color; // 0x24
     float fog; // 0x28
