+++ annotations/nocedit.exe/pseudocode/src/engine/3d.c/renderPolygonTextureWrappedPlaneMasked_FUN_00405b50.c
@@ -55,11 +55,10 @@ engine_3d_c_renderPolygonTextureWrappedPlaneMasked_FUN_00405b50(SMRGLHeaderPrimi
     pSVar2 = polygon_info + 1;
     for (iVar1 = 0; iVar1 < (polygon_info->base).count * 3; iVar1 = iVar1 + 3) {
       *piVar3 = (pSVar2->base).type + g_ProcessedVertexOffset;
-      g_RenderVertexBuffer[(pSVar2->base).type + g_ProcessedVertexOffset].u =
-           (float)(pSVar2->base).count;
+      g_RenderVertexBuffer[(pSVar2->base).type + g_ProcessedVertexOffset].u = (pSVar2->base).count;
       piVar3 = piVar3 + 1;
       g_RenderVertexBuffer[(pSVar2->base).type + g_ProcessedVertexOffset].v =
-           (float)(pSVar2->surface_normal).A;
+           (pSVar2->surface_normal).A;
       pSVar2 = (SMRGLHeaderPrimitive *)&(pSVar2->surface_normal).B;
     }
     engine_clipper_c_clipAndRasterize_FUN_004371b0
